integrate mini-game
