+++ README.md
@@ -128,6 +128,8 @@ Hintergrundbild: [Bild von rawpixel.com](https://de.freepik.com/fotos-kostenlos/
 
 Farbkleckse: [Bild von starline](https://de.freepik.com/vektoren-kostenlos/isolierter-aquarell-spritzerfleck-bunter-satz-von-acht_9728054.htm#query=klecks&position=4&from_view=keyword&track=sph) auf Freepik
 
+Snake-Icon: [Snake icons created by apien - Flaticon](https://www.flaticon.com/free-icons/snake)
+
 ## Entwicklung
 
 Installiere node.js und git. Um das Repo dann zum Laufen zu bringen, benötigt es folgende Schritte:
