+++ lib/commands/init.ts
@@ -117,6 +117,10 @@ export async function initClient(core: Core) {
               ws.analyze.useJava++
               continue
             }
+            if (entry.event == 'play_snake') {
+              ws.analyze.playSnake++
+              continue
+            }
             const publish = /publish_custom_quest_(.+)/.exec(entry.event)
             if (publish) {
               ws.analyze.published.push({
