+++ lib/state/types.ts
@@ -171,6 +171,7 @@ export interface Analyze {
   showStructogram: number
   usePersist: number
   useJava: number
+  playSnake: number
   customQuests: { [key: string]: { start: number; complete: number } }
   legacy: { [key: string]: { count: number } }
   appearance: { [key: string]: { count: number } }
